+++ src/components/Loader/index.ts
@@ -1,0 +1,9 @@
+import Loader from './Loader.vue'
+
+import { App } from 'vue'
+
+Loader.install = (app: App) => {
+  app.component('Loader', Loader)
+}
+
+export default Loader