+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 - Add more padding to the popover - [#83](https://github.com/PrefectHQ/miter-design/pull/83)
+- Export Loader component - [#82](https://github.com/PrefectHQ/miter-design/pull/82)
+- Improve Spacing CSS - [#84](https://github.com/PrefectHQ/miter-design/pull/84)
+- Update License to Apache 2.0 - [86](https://github.com/PrefectHQ/miter-design/pull/86)
 - Fix broken tests - [#81](https://github.com/PrefectHQ/miter-design/pull/81)
 - Standardize icon size classes - [#80](https://github.com/PrefectHQ/miter-design/pull/80)
 - Add loader component - [#79](https://github.com/PrefectHQ/miter-design/pull/79)